Support add-relation for remote services
