+++ api/client.go
@@ -143,20 +143,6 @@ func (c *Client) ServiceGet(service string) (*params.ServiceGetResults, error) {
 	return &results, err
 }
 
-// AddRelation adds a relation between the specified endpoints and returns the relation info.
-func (c *Client) AddRelation(endpoints ...string) (*params.AddRelationResults, error) {
-	var addRelRes params.AddRelationResults
-	params := params.AddRelation{Endpoints: endpoints}
-	err := c.facade.FacadeCall("AddRelation", params, &addRelRes)
-	return &addRelRes, err
-}
-
-// DestroyRelation removes the relation between the specified endpoints.
-func (c *Client) DestroyRelation(endpoints ...string) error {
-	params := params.DestroyRelation{Endpoints: endpoints}
-	return c.facade.FacadeCall("DestroyRelation", params, nil)
-}
-
 // ServiceCharmRelations returns the service's charms relation names.
 func (c *Client) ServiceCharmRelations(service string) ([]string, error) {
 	var results params.ServiceCharmRelationsResults
