+++ apiserver/client/client_test.go
@@ -1626,187 +1626,6 @@ func (s *clientRepoSuite) TestClientServiceDeployWithNetworks(c *gc.C) {
 	c.Assert(err, gc.ErrorMatches, "use of --networks is deprecated. Please use spaces")
 }
 
-func (s *clientSuite) checkEndpoints(c *gc.C, endpoints map[string]charm.Relation) {
-	c.Assert(endpoints["wordpress"], gc.DeepEquals, charm.Relation{
-		Name:      "db",
-		Role:      charm.RelationRole("requirer"),
-		Interface: "mysql",
-		Optional:  false,
-		Limit:     1,
-		Scope:     charm.RelationScope("global"),
-	})
-	c.Assert(endpoints["mysql"], gc.DeepEquals, charm.Relation{
-		Name:      "server",
-		Role:      charm.RelationRole("provider"),
-		Interface: "mysql",
-		Optional:  false,
-		Limit:     0,
-		Scope:     charm.RelationScope("global"),
-	})
-}
-
-func (s *clientSuite) assertAddRelation(c *gc.C, endpoints []string) {
-	s.setUpScenario(c)
-	res, err := s.APIState.Client().AddRelation(endpoints...)
-	c.Assert(err, jc.ErrorIsNil)
-	s.checkEndpoints(c, res.Endpoints)
-	// Show that the relation was added.
-	wpSvc, err := s.State.Service("wordpress")
-	c.Assert(err, jc.ErrorIsNil)
-	rels, err := wpSvc.Relations()
-	// There are 2 relations - the logging-wordpress one set up in the
-	// scenario and the one created in this test.
-	c.Assert(len(rels), gc.Equals, 2)
-	mySvc, err := s.State.Service("mysql")
-	c.Assert(err, jc.ErrorIsNil)
-	rels, err = mySvc.Relations()
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(len(rels), gc.Equals, 1)
-}
-
-func (s *clientSuite) TestSuccessfullyAddRelation(c *gc.C) {
-	endpoints := []string{"wordpress", "mysql"}
-	s.assertAddRelation(c, endpoints)
-}
-
-func (s *clientSuite) TestBlockDestroyAddRelation(c *gc.C) {
-	s.BlockDestroyEnvironment(c, "TestBlockDestroyAddRelation")
-	s.assertAddRelation(c, []string{"wordpress", "mysql"})
-}
-func (s *clientSuite) TestBlockRemoveAddRelation(c *gc.C) {
-	s.BlockRemoveObject(c, "TestBlockRemoveAddRelation")
-	s.assertAddRelation(c, []string{"wordpress", "mysql"})
-}
-
-func (s *clientSuite) TestBlockChangesAddRelation(c *gc.C) {
-	s.setUpScenario(c)
-	s.BlockAllChanges(c, "TestBlockChangesAddRelation")
-	_, err := s.APIState.Client().AddRelation([]string{"wordpress", "mysql"}...)
-	s.AssertBlocked(c, err, "TestBlockChangesAddRelation")
-}
-
-func (s *clientSuite) TestSuccessfullyAddRelationSwapped(c *gc.C) {
-	// Show that the order of the services listed in the AddRelation call
-	// does not matter.  This is a repeat of the previous test with the service
-	// names swapped.
-	endpoints := []string{"mysql", "wordpress"}
-	s.assertAddRelation(c, endpoints)
-}
-
-func (s *clientSuite) TestCallWithOnlyOneEndpoint(c *gc.C) {
-	s.setUpScenario(c)
-	endpoints := []string{"wordpress"}
-	_, err := s.APIState.Client().AddRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, "no relations found")
-}
-
-func (s *clientSuite) TestCallWithOneEndpointTooMany(c *gc.C) {
-	s.setUpScenario(c)
-	endpoints := []string{"wordpress", "mysql", "logging"}
-	_, err := s.APIState.Client().AddRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, "cannot relate 3 endpoints")
-}
-
-func (s *clientSuite) TestAddAlreadyAddedRelation(c *gc.C) {
-	s.setUpScenario(c)
-	// Add a relation between wordpress and mysql.
-	endpoints := []string{"wordpress", "mysql"}
-	eps, err := s.State.InferEndpoints(endpoints...)
-	c.Assert(err, jc.ErrorIsNil)
-	_, err = s.State.AddRelation(eps...)
-	c.Assert(err, jc.ErrorIsNil)
-	// And try to add it again.
-	_, err = s.APIState.Client().AddRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, `cannot add relation "wordpress:db mysql:server": relation already exists`)
-}
-
-func (s *clientSuite) setupRelationScenario(c *gc.C, endpoints []string) *state.Relation {
-	s.setUpScenario(c)
-	// Add a relation between the endpoints.
-	eps, err := s.State.InferEndpoints(endpoints...)
-	c.Assert(err, jc.ErrorIsNil)
-	relation, err := s.State.AddRelation(eps...)
-	c.Assert(err, jc.ErrorIsNil)
-	return relation
-}
-
-func (s *clientSuite) assertDestroyRelation(c *gc.C, endpoints []string) {
-	s.assertDestroyRelationSuccess(
-		c,
-		s.setupRelationScenario(c, endpoints),
-		endpoints)
-}
-
-func (s *clientSuite) assertDestroyRelationSuccess(c *gc.C, relation *state.Relation, endpoints []string) {
-	err := s.APIState.Client().DestroyRelation(endpoints...)
-	c.Assert(err, jc.ErrorIsNil)
-	// Show that the relation was removed.
-	c.Assert(relation.Refresh(), jc.Satisfies, errors.IsNotFound)
-}
-
-func (s *clientSuite) TestSuccessfulDestroyRelation(c *gc.C) {
-	endpoints := []string{"wordpress", "mysql"}
-	s.assertDestroyRelation(c, endpoints)
-}
-
-func (s *clientSuite) TestSuccessfullyDestroyRelationSwapped(c *gc.C) {
-	// Show that the order of the services listed in the DestroyRelation call
-	// does not matter.  This is a repeat of the previous test with the service
-	// names swapped.
-	endpoints := []string{"mysql", "wordpress"}
-	s.assertDestroyRelation(c, endpoints)
-}
-
-func (s *clientSuite) TestNoRelation(c *gc.C) {
-	s.setUpScenario(c)
-	endpoints := []string{"wordpress", "mysql"}
-	err := s.APIState.Client().DestroyRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, `relation "wordpress:db mysql:server" not found`)
-}
-
-func (s *clientSuite) TestAttemptDestroyingNonExistentRelation(c *gc.C) {
-	s.setUpScenario(c)
-	s.AddTestingService(c, "riak", s.AddTestingCharm(c, "riak"))
-	endpoints := []string{"riak", "wordpress"}
-	err := s.APIState.Client().DestroyRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, "no relations found")
-}
-
-func (s *clientSuite) TestAttemptDestroyingWithOnlyOneEndpoint(c *gc.C) {
-	s.setUpScenario(c)
-	endpoints := []string{"wordpress"}
-	err := s.APIState.Client().DestroyRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, "no relations found")
-}
-
-func (s *clientSuite) TestAttemptDestroyingPeerRelation(c *gc.C) {
-	s.setUpScenario(c)
-	s.AddTestingService(c, "riak", s.AddTestingCharm(c, "riak"))
-
-	endpoints := []string{"riak:ring"}
-	err := s.APIState.Client().DestroyRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, `cannot destroy relation "riak:ring": is a peer relation`)
-}
-
-func (s *clientSuite) TestAttemptDestroyingAlreadyDestroyedRelation(c *gc.C) {
-	s.setUpScenario(c)
-
-	// Add a relation between wordpress and mysql.
-	eps, err := s.State.InferEndpoints("wordpress", "mysql")
-	c.Assert(err, jc.ErrorIsNil)
-	rel, err := s.State.AddRelation(eps...)
-	c.Assert(err, jc.ErrorIsNil)
-
-	endpoints := []string{"wordpress", "mysql"}
-	err = s.APIState.Client().DestroyRelation(endpoints...)
-	// Show that the relation was removed.
-	c.Assert(rel.Refresh(), jc.Satisfies, errors.IsNotFound)
-
-	// And try to destroy it again.
-	err = s.APIState.Client().DestroyRelation(endpoints...)
-	c.Assert(err, gc.ErrorMatches, `relation "wordpress:db mysql:server" not found`)
-}
-
 func (s *clientSuite) TestClientWatchAll(c *gc.C) {
 	// A very simple end-to-end test, because
 	// all the logic is tested elsewhere.
@@ -3033,31 +2852,6 @@ func (s *clientSuite) TestBlockDestroyDestroySubordinateUnits(c *gc.C) {
 	assertLife(c, logging0, state.Alive)
 
 	s.assertDestroySubordinateUnits(c, wordpress0, logging0)
-}
-
-func (s *clientSuite) TestBlockRemoveDestroyRelation(c *gc.C) {
-	endpoints := []string{"wordpress", "mysql"}
-	relation := s.setupRelationScenario(c, endpoints)
-	// block remove-objects
-	s.BlockRemoveObject(c, "TestBlockRemoveDestroyRelation")
-	err := s.APIState.Client().DestroyRelation(endpoints...)
-	s.AssertBlocked(c, err, "TestBlockRemoveDestroyRelation")
-	assertLife(c, relation, state.Alive)
-}
-
-func (s *clientSuite) TestBlockChangeDestroyRelation(c *gc.C) {
-	endpoints := []string{"wordpress", "mysql"}
-	relation := s.setupRelationScenario(c, endpoints)
-	s.BlockAllChanges(c, "TestBlockChangeDestroyRelation")
-	err := s.APIState.Client().DestroyRelation(endpoints...)
-	s.AssertBlocked(c, err, "TestBlockChangeDestroyRelation")
-	assertLife(c, relation, state.Alive)
-}
-
-func (s *clientSuite) TestBlockDestroyDestroyRelation(c *gc.C) {
-	s.BlockDestroyEnvironment(c, "TestBlockDestroyDestroyRelation")
-	endpoints := []string{"wordpress", "mysql"}
-	s.assertDestroyRelation(c, endpoints)
 }
 
 func (s *clientSuite) TestDestroyEnvironment(c *gc.C) {
